Fight

+++ include/inventory.hpp
@@ -29,13 +29,13 @@ public:
     return false;
   }
 
-  void remove(int index) {
+  void remove(size_t index) {
     if (!elements.empty() && index < elements.size()) {
       elements.erase(elements.cbegin() + index);
     }
   }
 
-  bool element_available(int index) {
+  bool element_available(size_t index) {
     if (!elements.empty() && index < elements.size()) {
       return true;
     }
